refactor: seperate dashboard logic to files , fix: change logic of task not done
